pyproject.toml
[project]
name = "discourse"
version = "0.1.0"
description = "Turn-based Claude conversation orchestrator"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.scripts]
discourse = "discourse.main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
